backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.11.9
PyPDF2==3.0.1
python-docx==1.1.2
aiofiles==24.1.0
torch
torchaudio

